toyota angle rate limits

+++ board/safety_declarations.h
@@ -57,7 +57,7 @@ typedef struct {
   const bool has_steer_req_tolerance;
 
   // angle cmd limits
-  const int angle_deg_to_can;
+  const float angle_deg_to_can;
   const struct lookup_t angle_rate_up_lookup;
   const struct lookup_t angle_rate_down_lookup;
 } SteeringLimits;
